+++ src/test/resources/db/element/element_create_mysql.sql
@@ -9,7 +9,7 @@ CREATE TABLE IF NOT EXISTS elements
     element_id      bigint     NOT NULL COMMENT 'the internal unique primary index',
     formula_id      bigint     NOT NULL COMMENT 'each element can only be used for one formula',
     order_nbr       bigint     NOT NULL,
-    element_type_id bigint     NOT NULL,
+    element_type_id smallint   NOT NULL,
     user_id         bigint DEFAULT NULL,
     ref_id          bigint DEFAULT NULL COMMENT 'either a term,verb or formula id',
     resolved_text   varchar(255) DEFAULT NULL
