+++ src/test/resources/db/log/change_create.sql
@@ -11,7 +11,7 @@ CREATE TABLE IF NOT EXISTS changes
     user_id          bigint     NOT NULL,
     change_action_id smallint   NOT NULL,
     row_id           bigint DEFAULT NULL,
-    change_field_id  bigint     NOT NULL,
+    change_field_id  smallint   NOT NULL,
     old_value        text   DEFAULT NULL,
     new_value        text   DEFAULT NULL,
     old_id           bigint DEFAULT NULL,
