+++ src/test/resources/db/log/change_table_create_mysql.sql
@@ -6,7 +6,7 @@
 
 CREATE TABLE IF NOT EXISTS change_tables
 (
-    change_table_id   bigint           NOT NULL COMMENT 'the internal unique primary index',
+    change_table_id   smallint         NOT NULL COMMENT 'the internal unique primary index',
     change_table_name varchar(255)     NOT NULL COMMENT 'the real name',
     code_id           varchar(255) DEFAULT NULL COMMENT 'with this field tables can be combined in case of renaming',
     description       text         DEFAULT NULL COMMENT 'the user readable name'
